[rpc-loadgen] break down rpc_command_processor
